@@ -11,12 +11,13 @@
     },
     "devDependencies": {
         "@tailwindcss/forms": "^0.2.1",
+        "@vue/compiler-sfc": "^3.1.1",
         "alpinejs": "^2.7.3",
         "autoprefixer": "^10.1.0",
         "axios": "^0.21",
         "bootstrap": "^4.6.0",
         "jquery": "^3.6",
-        "laravel-mix": "^6.0.6",
+        "laravel-mix": "^6.0.0-beta.17",
         "lodash": "^4.17.19",
         "popper.js": "^1.16.1",
         "postcss": "^8.2.1",
@@ -25,8 +26,8 @@
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
         "tailwindcss": "^2.0.2",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.9.7",
+        "vue": "^3.1.1",
+        "vue-loader": "^16.2.0",
         "vue-template-compiler": "^2.6.12"
     }
 }
