@@ -1,23 +1,9 @@
 <?php
 
 use App\Http\Controllers\PlantsController;
+use App\Http\Controllers\PublicPlantsController;
 use App\Http\Controllers\TreeController;
 use Illuminate\Support\Facades\Route;
-
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
-
-Route::get('/', function () {
-    return view('welcome');
-});
 
 Route::resource('tree', TreeController::class);
 
@@ -27,7 +13,7 @@ Route::get('/admin', [App\Http\Controllers\AdminController::class, 'index'])->na
 Route::get('/admin/plants/delete/{plant}', [PlantsController::class, 'destroy'])->name('plants.delete');
 Route::resource('/admin/plants', PlantsController::class);
 
-
+Route::resource('/', PublicPlantsController::class);
 
 //Route::get('/dashboard', function () {
 //    return view('dashboard');
